добавляет корректировки
